% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixModels.R
\name{clusterMembership}
\alias{clusterMembership}
\title{Helper function to assing initial cluster membership to saccades.}
\usage{
clusterMembership(Saccades, k)
}
\arguments{
\item{Saccades}{Vector of saccade duration values}

\item{k}{Number of clusters}
}
\value{
returns vector of factors of initial cluster membership.
}
\description{
Helper function to assing initial cluster membership to saccades.
}
