% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixModels.R
\name{clusterProbabilities}
\alias{clusterProbabilities}
\title{Helper function to assing cluster membership probabilities to saccades.}
\usage{
clusterProbabilities(Saccades, k)
}
\arguments{
\item{Saccades}{Vector of saccade duration values}

\item{k}{Number of clusters}
}
\value{
Matrix with k columns of probabilites
}
\description{
Helper function to assing cluster membership probabilities to saccades.
}
