% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixModels.R
\name{fitMixModels}
\alias{fitMixModels}
\title{Function that uses the flexmix package to fit the mixture models.}
\usage{
fitMixModels(
  Saccades,
  type = c("Group", "Individual"),
  cluster = c("Random", "Membership", "Propabilities")
)
}
\arguments{
\item{Saccades}{A dataframe contaning the saccades duration and participant labels. Columnames must be Duration and PP respectively.}

\item{type}{Fit mixture models on the group as a whole or on the individual participants. Must be 'Group' or 'Individual'.}

\item{cluster}{Initial cluster assigments. Can be random, assignment into membership or probabilites.}
}
\value{
Returns a list containg all the fitted mixture models.
}
\description{
Function that uses the flexmix package to fit the mixture models.
}
