% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useGazepath.R
\name{useGazepath}
\alias{useGazepath}
\title{gazepath}
\usage{
useGazepath(
  PP,
  dirIn,
  dirResults = dirIn,
  filename,
  height_px = 1024,
  height_mm = 270,
  width_px = 1280,
  width_mm = 330,
  thres_dur = 100
)
}
\arguments{
\item{PP}{Vector of participants numbers that corresponds with the raw eye-tracking data csv files.}

\item{dirIn}{The directory to the raw eye-tracking data csv files}

\item{dirResults}{The directory to the saccades and fixations.}

\item{filename}{The filename of the results}

\item{height_px}{Height of the stimuli in pixels. This is a gazepath argument. The defeault is 1024.}

\item{height_mm}{Height of the stimuli in mm. This is a gazepath argument. The defeault is 270.}

\item{width_px}{Width of the stimuli in pixels. This is a gazepath argument. The defeault is 1280.}

\item{width_mm}{Width of the stimuli in mm. This is a gazepath argument. The defeault is 330.}

\item{thres_dur}{The duration threshold. This is a gazepath argument. The defeault is 100.}
}
\value{
Saves a .csv file to the results directory
}
\description{
Uses the gazepath function on all the raw eye-movement files in a folder.
These files have to be csv files with the following column names: LEFT_GAZE_X, LEFT_GAZE_Y, RIGHT_GAZE_X, RIGHT_GAZE_Y, Trial and Distance.
}
